:shower: メソッドの集約
