+++ lib/sakamichi_scraper/hinatazaka.rb
@@ -1,4 +1,4 @@
-require "sakamichi_scraper/base"
+require_relative "base"
 
 module SakamichiScraper
   class Hinatazaka < Base
@@ -33,7 +33,7 @@ module SakamichiScraper
     def picture_in_newest_article
       newest_article_url = article_urls_from_list_page(blog_top_page).first
       article_html = get_content(newest_article_url)
-      image_urls = image_urls_from_article_url(article_html)
+      image_urls = image_urls_from_article_url(article_html, "div.c-blog-article__text")
 
       mkdir_today_file_path unless Dir.exist?("img/#{exec_date}")
       download_images_from_url_list(image_urls)
@@ -45,25 +45,6 @@ module SakamichiScraper
       [].tap do |array|
         Nokogiri.parse(html, nil, nil).css(".p-blog-top__list > li").each do |c|
           array << "#{@home_page}#{c.css("a")[0][:href]}".match(/(.*)\?.*$/)[1]
-        end
-      end
-    end
-
-    def image_urls_from_article_url(article_html)
-      [].tap do |url|
-        Nokogiri::HTML.parse(article_html, nil, nil).css("div.c-blog-article__text img").each do |c|
-          url << c.attribute("src").value
-        end
-      end
-    end
-
-    def download_images_from_url_list(image_urls)
-      image_urls.each do |image_url|
-        dest_image_path = "img/#{exec_date}/#{image_url[%r([^/]+$)]}"
-        File.open(dest_image_path, "w") do |pass|
-          URI.parse(image_url).open do |img|
-            pass.write(img.read)
-          end
         end
       end
     end
