+++ lib/sakamichi_scraper/sakurazaka.rb
@@ -1,5 +1,4 @@
-require "sakamichi_scraper/base"
-require "open-uri"
+require_relative "base"
 
 module SakamichiScraper
   class Sakurazaka < Base
@@ -30,7 +29,7 @@ module SakamichiScraper
     def picture_in_newest_article
       newest_article_url = article_urls_from_list_page(blog_list_page).first
       article_html = get_content(newest_article_url)
-      image_urls = image_urls_from_article_url(article_html)
+      image_urls = image_urls_from_article_url(article_html, "div.box-article")
 
       mkdir_today_file_path unless Dir.exist?("img/#{exec_date}")
       download_images_from_url_list(image_urls)
@@ -42,28 +41,6 @@ module SakamichiScraper
       [].tap do |array|
         Nokogiri.parse(html, nil, nil).css(".com-blog-part.box4.fxpc > li").each do |c|
           array << "#{@home_page}#{c.css("a")[0][:href]}"
-        end
-      end
-    end
-
-    def image_urls_from_article_url(article_html)
-      [].tap do |url|
-        Nokogiri.parse(article_html, nil, nil).css("div.box-article img").each do |c|
-          image_url = c.attribute("src").value
-          next if exclude_img_path(@group_name).include?(image_url)
-
-          url << "#{@home_page}#{image_url}"
-        end
-      end
-    end
-
-    def download_images_from_url_list(image_urls)
-      image_urls.each do |image_url|
-        dest_image_path = "img/#{exec_date}/#{image_url[%r([^/]+$)]}"
-        File.open(dest_image_path, "w") do |pass|
-          URI.parse(image_url).open do |img|
-            pass.write(img.read)
-          end
         end
       end
     end
